Create mouse look and first person shooter camera component, cleaned regular component
